+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/valibot
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [65243ce]
+  - ai@4.2.8
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [e14c066]
+  - ai@4.2.7
+
 ## 0.1.6
 
 ### Patch Changes